refactor: src/lib/ → utils/ 분리
- lib/: 외부 라이브러리 래퍼만 유지 (dayjs, supabase-server)
- utils/: 유틸·헬퍼 함수 (utils, voterId, menu-policy)
