+++ src/components/@shared/layout/SiteHeader/SiteHeader.styles.ts
@@ -1,4 +1,4 @@
-import {cn} from '@/lib/utils';
+import {cn} from '@/utils/utils';
 
 export const headerClass = (scrolled: boolean, menuOpen: boolean) =>
   cn(
