+++ src/components/home/HeroStatus/hero-status.ts
@@ -3,8 +3,8 @@ import type {ElementType} from 'react';
 import {Clock, Moon, Sun, Utensils} from 'lucide-react';
 
 import dayjs from '@/lib/dayjs';
-import {CAFETERIA_CLOSE_MIN, CAFETERIA_LABEL, CAFETERIA_OPEN_MIN} from '@/lib/menu-policy';
-import {formatYYYYMMDD} from '@/lib/utils';
+import {CAFETERIA_CLOSE_MIN, CAFETERIA_LABEL, CAFETERIA_OPEN_MIN} from '@/utils/menu-policy';
+import {formatYYYYMMDD} from '@/utils/utils';
 import {MenuType} from '@/types/menu';
 
 export type HeroStatusState = {icon: ElementType; text: string};
