+++ src/components/menu/MenuBoard/MenuBoard.styles.ts
@@ -1,4 +1,4 @@
-import {cn} from '@/lib/utils';
+import {cn} from '@/utils/utils';
 
 export const todayButtonClass = (isActive: boolean) =>
   cn(
